t[HALX86][NTOS:KE] Addressing code reviews
