+++ hal/halx86/smp/smpp.h
@@ -7,6 +7,3 @@
 
 VOID
 HalpInitializeAPStub(PVOID APStubLocation);
-
-VOID
-HalpInitializeAPPageTables(PVOID APStubLocation);
