+++ ntoskrnl/ntos.cmake
@@ -186,6 +186,7 @@ list(APPEND SOURCE
     ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/gmutex.c
     ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/ipi.c
     ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/krnlinit.c
+    ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/mp.c
     ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/mutex.c
     ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/procobj.c
     ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/profobj.c
@@ -309,7 +310,6 @@ if(ARCH STREQUAL "i386")
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/irqobj.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/kiinit.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/ldt.c
-        ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/mp.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/mtrr.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/patpge.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/i386/thrdini.c
@@ -340,7 +340,6 @@ elseif(ARCH STREQUAL "amd64")
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/amd64/irql.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/amd64/kiinit.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/amd64/krnlinit.c
-        ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/amd64/mp.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/amd64/spinlock.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/ke/amd64/thrdini.c
         ${REACTOS_SOURCE_DIR}/ntoskrnl/mm/amd64/init.c
